Code cleanup Moved noise library to wrappers package
